Handler permissions

+++ apps/utils/complex_rest.py
@@ -74,8 +74,21 @@ class MongoEngineComplexDataManagerPerUser(MongoEngineComplexDataManager):
 
 
 class ComplexRestHandler(ApiHandler):
+    perm_public = 'CRLUD'
+    perm_user = 'CRLUD'
+    perm_admin = 'CRLUD'
+
+    def check_permission(self, action):
+        user = self.get_current_user()
+        admin = self.is_admin_user()
+        if action in self.perm_public or (user and action in self.perm_user) or (admin and action in self.perm_admin):
+            pass # ok
+        else:
+            self.raise403()
+
     # CREATE /objs
     def post(self, *identifiers, **kwargs):
+        self.check_permission('C')
         data = self.get_request_data()
         obj = self.data_manager.create(identifiers, data)
         if obj:
@@ -89,12 +102,14 @@ class ComplexRestHandler(ApiHandler):
         if len(identifiers) == 1 and identifiers[0] == 'count':
             self.answer(self.data_manager.count())
         elif len(identifiers) == self.data_manager.expected_keys():
+            self.check_permission('R')
             obj = self.data_manager.read(identifiers)
             if obj:
                 self.answer(obj)
             else:
                 self.answer(obj)
         else:
+            self.check_permission('L')
             # FIXME pagination
             initial = int(self.get_argument('initial', default=0))
             amount = int(self.get_argument('amount', default=50))
@@ -105,6 +120,7 @@ class ComplexRestHandler(ApiHandler):
 
     # UPDATE /objs/:id
     def put(self, *identifiers, **kwargs):
+        self.check_permission('U')
         data = self.get_request_data()
         obj = self.data_manager.update(identifiers, data)
         if obj:
@@ -114,6 +130,7 @@ class ComplexRestHandler(ApiHandler):
 
     # DELETE /objs/:id
     def delete(self, *identifiers, **kwargs):
+        self.check_permission('D')
         obj = self.data_manager.delete(identifiers)
         if obj:
             self.answer(obj)
